Client.Authorize error handling
modifications to client.authorize:
* if statement on line 161 checks if body.token exists, but if it’s an
incorrect authentication it will throw an error that will crash a server
* an incorrect login previously threw an error, now it passes the error
through to a callback. The error is short-circuited to be true, as i’ve
experienced times when an error is not thrown on incorrect login
attempts, but the function does not enter the first if statement.